Make VexFlow project directory configurable for SCons.
This will allow others to place the VexFlow sources wherever they
please, while still being able to build the Tutti package. The path can
be specified via command line flag `--vex-dir`. If no path is specified,
the default is to assume the project is named "vexflow" and that it
resides in the same directory as the root of the Rondo project.