Add Sencha Cmd support for VexFlow builds.
VexFlow builds in the Tutti package are integrated with Sencha Cmd and
can be run using the Ant target "vexflow". Users should provide the path
to their local copy of the VexFlow project by creating a simple
`local.properties` file in the workspace root that specifies a path for
the variable `vexflow.dir`.