Fix error handling in Selenium tests, and add additional test options.
Selenium tests now accept `autoclose` and `slow` options. The autoclose
option is enabled by default, and will automatically quit the browser at
the end of each test session. If set to `false`, the browser will remain
open at the end of the session for further manual testing.

The slow option is disabled by default. When enabled, it automatically
injects a half-second delay before every action, for debugging purposes
when a test fails. If disabled, tests will run at full speed.

These options can be set on the grunt command line via a new `--human`
flag, that disables autoclose and enables slow running on top of the
options included in `--local`. This should allow for easier diagnosis
of failing tests at a "human" level on the local machine.